+++ src/estampes/tools/output.py
@@ -3,9 +3,11 @@
 The module provides some simple tools for common output operations.
 """
 import re
+import typing as tp
 from collections.abc import Sequence
 
 from estampes.base import ArgumentError
+from estampes.data.property import QBaseInfo
 
 
 def fortran_fmt_D(number_s: float | Sequence[float], fmt: str) -> str:
@@ -93,3 +95,140 @@ def fortran_fmt_D(number_s: float | Sequence[float], fmt: str) -> str:
         return ''.join(repr_s)
     else:
         raise TypeError('Unsupported representation of number_s')
+
+
+def pstruct_to_labels(qinfo: QBaseInfo,
+                      natoms: int | None = None,
+                      linearize: bool = False,
+                      sep_atom: str = ' ',
+                      sep_xyz: str = '') -> str | tuple[tp.Any, ...]:
+    """Construct sequence of labels from property structure information.
+
+    Constructs a sequence of labels following the dimensions information
+    or linearized, based on the property structure description.
+
+    Parameters
+    ----------
+    qinfo
+        Quantity information
+    natoms
+        Number of atoms, only relevant for properties dependent on atoms.
+    linearize
+        Linearize structure instead of keeping the assumed shape.
+    sep_atom
+        Separator character(s) to use between atoms or atom-coordinates.
+        If not separator is provided, at least one space is added
+        between atoms labels to ensure that they are readable.
+    sep_xyz
+        Separator to use between coordinates of different dimensions.
+
+    Returns
+    -------
+    str or tuple
+        Sequence of labels.  An empty string is returned for scalars.
+    """
+    def proc_atom_spec(struct: str, nat: int | None) -> tuple[tp.Any, ...]:
+        if nat is None:
+            raise ArgumentError('natoms', 'Missing number of atoms')
+        if nat <= 0:
+            raise ValueError(
+                'Number of atoms cannot be negative or null')
+        fmt_lab = f'{{:>{len(str(nat))}c}}'
+        if struct == 'atom':
+            return tuple((fmt_lab.format(i+1) for i in range(nat)))
+        elif struct == 'atomLT':
+            if not isinstance(sep_atom, str):
+                raise TypeError('sep_atom must be a string')
+            return tuple((fmt_lab.format(i+1) + sep_atom + fmt_lab.format(j+1)
+                          for i in range(nat) for j in range(i+1)))
+        else:
+            raise ValueError('Unrecognized atom-based structure')
+
+    xyz = ('x', 'y', 'z')
+    comp_xyz = {
+        'xyz': xyz,
+        'xyz2D': (lab1+lab2 for lab1 in xyz for lab2 in xyz),
+        'xyz3D': (lab1+lab2+lab3
+                  for lab1 in xyz for lab2 in xyz for lab3 in xyz),
+        'xyzLT': (lab1+lab2
+                  for i, lab1 in enumerate(xyz, start=1)
+                  for lab2 in xyz[:i]),
+        'xyzLC': (lab1+lab2+lab3
+                  for i, lab1 in enumerate(xyz, start=1)
+                  for j, lab2 in enumerate(xyz[:i], start=1)
+                  for lab3 in xyz[:j])
+    }
+
+    match (qinfo.form):
+        case ('scalar'):
+            return ''
+        case str():
+            if qinfo.form.startswith('atom'):
+                return proc_atom_spec(qinfo.form, natoms)
+            else:
+                return tuple(comp_xyz[qinfo.form])
+        case tuple() | list():
+            comps = []
+            is_xyz = []
+            for item in qinfo.form:
+                if isinstance(item, str):
+                    if item.startswith('atom'):
+                        comps.append(proc_atom_spec(item, natoms))
+                        is_xyz.append(False)
+                    else:
+                        comps.append(comp_xyz[item])
+                        is_xyz.append(True)
+                elif isinstance(item, (tuple, list)):
+                    # We assume there is only one max level of nesting
+                    # So at most a structure tuple(tuple(scalar))
+                    # which corresponds to most standard cases.
+                    for subitem in item:
+                        if subitem.startswith('atom'):
+                            comps.append(proc_atom_spec(subitem, natoms))
+                            is_xyz.append(False)
+                        else:
+                            comps.append(comp_xyz[subitem])
+                            is_xyz.append(True)
+            # We build list of components, now construct extended structure.
+            # We could do a generic but in practice we will have 2 or 3
+            # dimensions, so we make this explicit for simplicity.
+            if any(is_xyz):
+                if not isinstance(sep_xyz, str):
+                    raise TypeError('String expected as separator between '
+                                    + 'Cartesian components')
+            if not all(is_xyz):
+                if not isinstance(sep_atom, str):
+                    raise TypeError('String expected as separator between '
+                                    + 'atomic components')
+            if linearize:
+                match (len(comps)):
+                    case 2:
+                        sep = sep_xyz if is_xyz[0] else sep_atom
+                        return tuple((lab1+sep+lab2
+                                      for lab1 in comps[0]
+                                      for lab2 in comps[1]))
+                    case 3:
+                        sep1 = sep_xyz if is_xyz[0] else sep_atom
+                        sep2 = sep_xyz if is_xyz[1] else sep_atom
+                        return tuple((lab1+sep1+lab2+sep2+lab3
+                                      for lab1 in comps[0]
+                                      for lab2 in comps[1]
+                                      for lab3 in comps[2]))
+                    case _:
+                        raise ValueError('Unsupported component structure')
+            else:
+                match (len(comps)):
+                    case 2:
+                        sep = sep_xyz if is_xyz[0] else sep_atom
+                        return tuple(((lab1+sep+lab2
+                                       for lab1 in comps[0])
+                                      for lab2 in comps[1]))
+                    case 3:
+                        sep1 = sep_xyz if is_xyz[0] else sep_atom
+                        sep2 = sep_xyz if is_xyz[1] else sep_atom
+                        return tuple((((lab1+sep1+lab2+sep2+lab3
+                                        for lab1 in comps[0])
+                                       for lab2 in comps[1])
+                                      for lab3 in comps[2]))
+                    case _:
+                        raise ValueError('Unsupported component structure')